refactor: remove unused data files and components from the About section
