--- client/src/data/Crops/Crops.js
@@ -1,10 +1,0 @@
-export const cropTypes = [
-  { id: "corn", label: "Corn" },
-  { id: "wheat", label: "Wheat" },
-  { id: "rice", label: "Rice" },
-  { id: "soybeans", label: "Soybeans" },
-  { id: "potatoes", label: "Potatoes" },
-  { id: "tomatoes", label: "Tomatoes" },
-  { id: "cotton", label: "Cotton" },
-  { id: "sugarcane", label: "Sugarcane" },
-];
