--- client/src/data/Driver/Driver.js
@@ -1,8 +1,0 @@
-export const vehicleTypes = [
-  { value: "pickup", label: "Pickup Truck" },
-  { value: "small_truck", label: "Small Truck" },
-  { value: "medium_truck", label: "Medium Truck" },
-  { value: "large_truck", label: "Large Truck / Semi" },
-  { value: "refrigerated", label: "Refrigerated Truck" },
-  { value: "van", label: "Van" },
-];
