--- client/src/data/aboutUs/aboutUsCoreValues.js
@@ -1,26 +1,0 @@
-export const values = [
-  {
-    icon: "🌱",
-    title: "Sustainability",
-    description:
-      "Building technology that promotes environmentally responsible farming from day one.",
-  },
-  {
-    icon: "🤝",
-    title: "Community",
-    description:
-      "Creating connections between farmers, vendors, and transporters for a stronger agricultural ecosystem.",
-  },
-  {
-    icon: "💡",
-    title: "Innovation",
-    description:
-      "Applying cutting-edge technology to solve age-old problems in agriculture.",
-  },
-  {
-    icon: "⚖️",
-    title: "Fairness",
-    description:
-      "Developing transparent processes that ensure equitable outcomes for all platform participants.",
-  },
-];
