--- client/src/data/aboutUs/aboutUsMilestones.js
@@ -1,32 +1,0 @@
-export const milestones = [
-  {
-    date: "January 2023",
-    title: "Idea Conception",
-    description:
-      "CropMate began as a concept during an agricultural hackathon in Colombo.",
-  },
-  {
-    date: "March 2023",
-    title: "Seed Funding",
-    description:
-      "Secured initial funding from local angel investors who believe in our vision.",
-  },
-  {
-    date: "June 2023",
-    title: "Beta Launch",
-    description:
-      "Released our first prototype to 50 farmers and 10 vendors in Kandy district.",
-  },
-  {
-    date: "October 2023",
-    title: "Team Expansion",
-    description:
-      "Grew our team to include agricultural experts and additional developers.",
-  },
-  {
-    date: "January 2024",
-    title: "Official Launch",
-    description:
-      "Publicly launched CropMate platform with improved features based on beta feedback.",
-  },
-];
