--- client/src/data/aboutUs/aboutUsOurTeam.js
@@ -1,20 +1,0 @@
-export const teamMembers = [
-  {
-    name: "Anjali Patel",
-    role: "Founder & CEO",
-    bio: "Former agricultural consultant with a vision to transform Sri Lankan farming through accessible technology.",
-    image: "/src/assets/images/team/founder.webp",
-  },
-  {
-    name: "Rajiv Mendis",
-    role: "CTO & Co-founder",
-    bio: "Full-stack developer who left his tech job to help build the digital future of agriculture in his homeland.",
-    image: "/src/assets/images/team/cto.webp",
-  },
-  {
-    name: "Dharshana Silva",
-    role: "Agricultural Advisor",
-    bio: "Recent PhD graduate in Agricultural Sciences bringing fresh research and perspectives to sustainable farming.",
-    image: "/src/assets/images/team/specialist.webp",
-  },
-];
