--- client/src/data/aboutUs/aboutUsStats.js
@@ -1,6 +1,0 @@
-export const stats = [
-  { value: "50+", label: "Beta Users" },
-  { value: "3", label: "Districts Piloting" },
-  { value: "25%", label: "Test Group Yield Increase" },
-  { value: "2025", label: "Founded" },
-];
