+++ client/src/pages/About/OurValues.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { motion } from 'framer-motion'
 import SectionWrapper from '../../components/Shared/SectionWrapper'
 import SectionHeader from '../../components/Shared/SectionHeader'
-import { values } from '../../data/aboutUs/aboutUsCoreValues'
+import { values } from '../../constants'
 import { containerVariants, itemVariants } from '../../util/animations'
 
 const OurValues = () => {
